+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/IrContributionData.kt
@@ -9,15 +9,12 @@ import dev.zacsweers.metro.compiler.mapToSet
 import dev.zacsweers.metro.compiler.reportCompilerBug
 import org.jetbrains.kotlin.descriptors.Visibilities
 import org.jetbrains.kotlin.ir.declarations.IrClass
-import org.jetbrains.kotlin.ir.declarations.IrDeclarationWithVisibility
-import org.jetbrains.kotlin.ir.declarations.IrFile
-import org.jetbrains.kotlin.ir.declarations.moduleDescriptor
 import org.jetbrains.kotlin.ir.types.IrType
 import org.jetbrains.kotlin.ir.types.classOrFail
 import org.jetbrains.kotlin.ir.types.classOrNull
 import org.jetbrains.kotlin.ir.util.defaultType
 import org.jetbrains.kotlin.ir.util.file
-import org.jetbrains.kotlin.ir.util.getPackageFragment
+import org.jetbrains.kotlin.ir.util.fileOrNull
 import org.jetbrains.kotlin.ir.util.nestedClasses
 import org.jetbrains.kotlin.name.ClassId
 
@@ -49,13 +46,17 @@ internal class IrContributionData(private val metroContext: IrMetroContext) {
     addAll(findExternalBindingContainerContributions(scope))
   }
 
-  private fun findVisibleContributionClassesForScopeInHints(scope: Scope): Set<IrClass> {
+  fun findVisibleContributionClassesForScopeInHints(
+    scope: Scope,
+    includeNonFriendInternals: Boolean = false,
+  ): Set<IrClass> {
     val functionsInPackage = metroContext.referenceFunctions(Symbols.CallableIds.scopeHint(scope))
     val contributingClasses =
       functionsInPackage
         .filter {
           if (it.owner.visibility == Visibilities.Internal) {
-            it.owner.isVisibleAsInternal(it.owner.file)
+            includeNonFriendInternals ||
+              it.owner.fileOrNull?.let { file -> it.owner.isVisibleAsInternal(file) } ?: false
           } else {
             true
           }
@@ -141,14 +142,5 @@ internal class IrContributionData(private val metroContext: IrMetroContext) {
           }
         }
       }
-  }
-
-  // Copied from CheckerUtils.kt
-  private fun IrDeclarationWithVisibility.isVisibleAsInternal(file: IrFile): Boolean {
-    val referencedDeclarationPackageFragment = getPackageFragment()
-    val module = file.module
-    return module.descriptor.shouldSeeInternalsOf(
-      referencedDeclarationPackageFragment.moduleDescriptor
-    )
   }
 }